rename addraster command to add-raster
