+++ src/stactools/core/add_raster.py
